Added corner icon support to OnScreenDisplay
Primarily used for RetroAchievements support, this stores a set of icons (pointers to the texture cache) to display in a vertical list on the bottom right hand corner of the game window. These icons persist as long as they are in the set, and methods to add to, remove from, and clear out the set are exposed to the user.