Added event handling for Primed Achievement Events
When an achievement is Primed, then every criteria is met for its unlocking except for its Trigger. To denote this, the achievement's badge is displayed on the emulator window. HandleAchievementPrimedEvent thus adds the icon to OSD's set for displaying, and HandleAchievementUnprimedEvent removes it. When the game is closed, OSD's set is cleared.