Rearranged RetroAchievements startup process
Moved AchievementManager Init further down in the MainWindow constructor; its original position was because it had an impact on the contents of the menu bar, and this is no longer the case. On top of this, if achievements are enabled and there's an API token already in the config (meaning the emulator will attempt to verify credentials and "log in"), it opens the Achievements dialog as a way of notifying the player that the achievements are active.