RetroAchievements - Prevent Multithreaded Memory Reads
The RetroAchievements client now comes with a flag that forces rcheevos, and more importantly RAIntegration, to only perform memory reads and writes on calls to do_frame or do_idle. This ensures that as long as those are only called in the CPU thread, all memory calls can be done on that thread. This allows for two considerable simplifications to the AchievementManager code - MemoryVerifier is no longer necessary upon game start, and the memory clone needed by the dev tools for access can also be deleted. The latter in particular will greatly improve performance when running Dolphin with the dev tools.